consolidate UniProcExecutor
